+++ TerraformingMars/model/src/main/java/com/tm/CEP/CardPlayedEvent.java
@@ -13,7 +13,7 @@ public class CardPlayedEvent extends CEP {
 
     private CardTag cardTag;
 
-    public CardPlayedEvent(double playerId, long timestamp, int generation, CardTag cardTag) {
+    public CardPlayedEvent(long playerId, long timestamp, int generation, CardTag cardTag) {
         super(playerId, timestamp, generation);
         this.cardTag = cardTag;
     }
