+++ TerraformingMars/model/src/main/java/com/tm/CEP/TilePlayedEvent.java
@@ -12,7 +12,7 @@ public class TilePlayedEvent extends CEP {
 
     private TileType tileType;
 
-    public TilePlayedEvent(double playerId, long timestamp, int generation, TileType tileType) {
+    public TilePlayedEvent(long playerId, long timestamp, int generation, TileType tileType) {
         super(playerId, timestamp, generation);
         this.tileType = tileType;
     }
